Move getFullEndpointApiUrl
