+++ src/frontend/packages/store/src/endpoint-utils.ts
@@ -1,0 +1,6 @@
+import { EndpointModel } from './types/endpoint.types';
+
+export function getFullEndpointApiUrl(endpoint: EndpointModel) {
+  return endpoint && endpoint.api_endpoint ?
+    `${endpoint.api_endpoint.Scheme}://${endpoint.api_endpoint.Host}${endpoint.api_endpoint.Path}` : 'Unknown';
+}
